+++ lib/microstrain_mag_cal_apply/composed_coefficients.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <microstrain_mag_cal/calibration.hpp>
+
+
+namespace microstrain_mag_cal
+{
+    FitResult composeCorrections(const FitResult &old_fit, const FitResult &new_fit);
+};